refactor: move rpc and provider setup into iroh
