+++ iroh/src/lib.rs
@@ -3,6 +3,9 @@
 pub use iroh_bytes as bytes;
 pub use iroh_net as net;
 
+pub mod node;
+pub mod rpc_protocol;
+
 #[cfg(feature = "cli")]
 pub mod config;
 #[cfg(feature = "cli")]
